test-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe --extra=test --output-file=test-requirements.txt
#
ase==3.22.1
    # via pysisyphus (setup.py)
attrs==21.4.0
    # via pytest
autograd==1.3
    # via pysisyphus (setup.py)
bcrypt==3.2.0
    # via paramiko
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
click==8.0.3
    # via distributed
cloudpickle==2.0.0
    # via
    #   dask
    #   distributed
coverage[toml]==6.2
    # via pytest-cov
cryptography==36.0.1
    # via paramiko
cycler==0.11.0
    # via matplotlib
dask==2022.01.0
    # via
    #   distributed
    #   pysisyphus (setup.py)
distributed==2022.1.0
    # via pysisyphus (setup.py)
fabric==2.6.0
    # via pysisyphus (setup.py)
fonttools==4.28.5
    # via matplotlib
fsspec==2022.1.0
    # via dask
future==0.18.2
    # via autograd
h5py==3.6.0
    # via
    #   pyscf
    #   pysisyphus (setup.py)
heapdict==1.0.1
    # via zict
iniconfig==1.1.1
    # via pytest
invoke==1.6.0
    # via fabric
jinja2==3.0.3
    # via
    #   distributed
    #   pysisyphus (setup.py)
joblib==1.1.0
    # via scikit-learn
kiwisolver==1.3.2
    # via matplotlib
locket==0.2.1
    # via partd
markupsafe==2.0.1
    # via jinja2
matplotlib==3.5.1
    # via
    #   ase
    #   pysisyphus (setup.py)
mpmath==1.2.1
    # via sympy
msgpack==1.0.3
    # via distributed
natsort==8.0.2
    # via pysisyphus (setup.py)
numpy==1.22.1
    # via
    #   ase
    #   autograd
    #   h5py
    #   matplotlib
    #   pyscf
    #   pysisyphus (setup.py)
    #   qcelemental
    #   rmsd
    #   scikit-learn
    #   scipy
packaging==21.3
    # via
    #   dask
    #   distributed
    #   matplotlib
    #   pint
    #   pytest
paramiko==2.9.2
    # via fabric
partd==1.2.0
    # via dask
pathlib2==2.3.6
    # via fabric
pillow==9.0.0
    # via matplotlib
pint==0.18
    # via qcelemental
pluggy==1.0.0
    # via pytest
psutil==5.9.0
    # via
    #   distributed
    #   pysisyphus (setup.py)
    #   qcengine
py==1.11.0
    # via pytest
py-cpuinfo==8.0.0
    # via qcengine
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via
    #   qcelemental
    #   qcengine
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pyscf==2.0.1
    # via pysisyphus (setup.py)
pytest==6.2.5
    # via
    #   pysisyphus (setup.py)
    #   pytest-cov
pytest-cov==3.0.0
    # via pysisyphus (setup.py)
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via
    #   dask
    #   distributed
    #   pysisyphus (setup.py)
    #   qcengine
qcelemental==0.24.0
    # via qcengine
qcengine==0.21.0
    # via pysisyphus (setup.py)
rmsd==1.4
    # via pysisyphus (setup.py)
scikit-learn==1.0.2
    # via pysisyphus (setup.py)
scipy==1.7.3
    # via
    #   ase
    #   pyscf
    #   pysisyphus (setup.py)
    #   rmsd
    #   scikit-learn
six==1.16.0
    # via
    #   bcrypt
    #   pathlib2
    #   python-dateutil
sortedcontainers==2.4.0
    # via distributed
sympy==1.9
    # via pysisyphus (setup.py)
tblib==1.7.0
    # via distributed
threadpoolctl==3.0.0
    # via scikit-learn
toml==0.10.2
    # via pytest
tomli==2.0.0
    # via coverage
toolz==0.11.2
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via distributed
typing-extensions==4.0.1
    # via pydantic
zict==2.0.0
    # via distributed

# The following packages are considered to be unsafe in a requirements file:
setuptools==60.5.0
    # via distributed
